update docker compose and property apis

+++ src/property/property.service.ts
@@ -11,29 +11,34 @@ export class PropertyService {
   constructor(
     @InjectRepository(Property) private propertyRepo: Repository<Property>,
   ) {}
-  async findAll(query: PaginationDto) {
+  async findAll(query: PaginationDto, userId: number) {
     return await this.propertyRepo.find({
       skip: query.skip,
       take: query.limit,
+      where: {
+        user: { id: userId },
+      },
     });
   }
 
-  async findOne(id: number) {
-    const prop = await this.propertyRepo.findOne({ where: { id } });
+  async findOne(id: number, userId: number) {
+    const prop = await this.propertyRepo.findOne({
+      where: { id, user: { id: userId } },
+    });
 
     if (!prop) throw new NotFoundException();
     return prop;
   }
 
-  async create(dto: CreatePropertyDto) {
-    return await this.propertyRepo.save(dto);
+  async create(dto: CreatePropertyDto, userId: number) {
+    return await this.propertyRepo.save({ ...dto, user: { id: userId } });
   }
 
-  async update(id: number, dto: UpdatePropertyDto) {
-    return await this.propertyRepo.update({ id }, dto);
+  async update(id: number, dto: UpdatePropertyDto, userId: number) {
+    return await this.propertyRepo.update({ id, user: { id: userId } }, dto);
   }
 
-  async delete(id: number) {
-    return await this.propertyRepo.delete({ id });
+  async delete(id: number, userId: number) {
+    return await this.propertyRepo.delete({ id, user: { id: userId } });
   }
 }
